Add caching of registration data to speed things up
We're aggregating survey responses on a per-wiki basis, and
would grab registration data for each wiki. This works well when
you have a handful of wikis, now that we're on hundreds of them
it's no longer a good idea. This update caches a filtered view
of the registration data, then queries that cached view. This
leads to much faster querying, but we're still looking at a bit
of wall time for each wiki. We therefore also set a two hour
timeout on completing the notebook.